Fixing typos and making the prompt scroll better
The prompt game output terminal won't overflow on itself and  look weird when resized, now every line will be at it's full length and not all... "wrapped" together as they used to be. for example if the width was 10 characters this would appear when a line of 20 characters was printed
```
0123456789
0123456789
```
and if the pane was expanded to be more than 10 character wide it would stay that way
now no matter what amount of character is on a line it will you will always be able to scroll to the end of it without the size of the window from when it was printed affecting anything,

also fixed a typo on the windows exe paths which were typed winows instead of windows thanks to silverfeelin for noticing and telling me about it.